+++ src/main/java/com/app/kokonut/auth/jwt/been/JwtTokenProvider.java
@@ -43,8 +43,8 @@ public class JwtTokenProvider {
     private static final String AUTHORITIES_KEY = "Authorization";
 
 //    private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 30;              // 30분
-    private static final long REFRESH_TOKEN_EXPIRE_TIME = 1000 * 60 * 60 * 24 * 7;    // 7일
-    private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 5;      // 5분
+    private static final long REFRESH_TOKEN_EXPIRE_TIME = (long)1000 * 60 * 60 * 24 * 7;    // 7일
+    private static final long ACCESS_TOKEN_EXPIRE_TIME = (long)1000 * 60 * 5;      // 5분
 //    private static final long REFRESH_TOKEN_EXPIRE_TIME = 1000 * 60 * 10;    // 10분
 
     private Key key;
